Rewrite computed properties with new syntax to suppress deprecation warnings
Ember.js 1.12 introduced an improved syntax[1] for computed properties with
a setter:

```
DEPRECATION: Using the same function as getter and setter is deprecated. See http://emberjs.com/deprecations/v1.x/#toc_computed-properties-with-a-shared-getter-and-setter for more details.
        at ComputedProperty (http://localhost:7357/static-assets/vendor.js:20233)
        at computed (http://localhost:7357/static-assets/vendor.js:20694)
        at http://localhost:7357/static-assets/vendor.js:79924
        at http://localhost:7357/static-assets/vendor.js:80572
```

Use new getter and setter syntax instead of checking arguments.length and add Ember version 1.13.2.

Caveat: this change is not compatible with Ember version below 1.12.0. In Ember 1.11.x the
        "new-computed-syntax" feature has to be explicitly enabled to make this work,
        see https://github.com/rwjblue/ember.js/commit/9ee6d7c5db965ab0408fa77dc6e624d7ad322431.